Victory Condition, Stat Saving.

+++ Assets/Scripts/Architecture/Persistence/TimeTrialData.cs
@@ -5,7 +5,7 @@ namespace Architecture.Persistence
     [Serializable]
     public class TimeTrialData
     {
-        public int BestTime = 0;
+        public float BestTime = 0;
         public int TimesPlayed = 0;
         public int PowerUpsCollected = 0;
         public MedalType BestMedalReached = MedalType.None;
